move clipp/parse logic into console project
