+++ common/common.vcxitems
@@ -24,8 +24,8 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)accel-sigmoid.hpp" />
     <ClInclude Include="$(MSBuildThisFileDirectory)accel-error.hpp" />
     <ClInclude Include="$(MSBuildThisFileDirectory)rawaccel-io.hpp" />
-    <ClInclude Include="$(MSBuildThisFileDirectory)rawaccel-userspace.hpp" />
     <ClInclude Include="$(MSBuildThisFileDirectory)rawaccel.hpp" />
+    <ClInclude Include="$(MSBuildThisFileDirectory)tagged-union-single.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)x64-util.hpp" />
     <ClInclude Include="$(MSBuildThisFileDirectory)vec2.h" />
   </ItemGroup>
