+++ common/tagged-union-single.h
