+++ console/console.vcxproj
@@ -91,6 +91,9 @@
   <ItemGroup>
     <ClCompile Include="console.cpp" />
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="parse.hpp" />
+  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
