+++ console/external/clipp.h
